Fix logic and relocate code
