+++ auto_build/oss_fuzz_templates.py
